+++ src/main/java/com/codecool/navymanager/service/GunService.java
@@ -2,15 +2,23 @@ package com.codecool.navymanager.service;
 
 
 import com.codecool.navymanager.dto.GunDto;
+import com.codecool.navymanager.entity.Country;
+import com.codecool.navymanager.entity.Gun;
 import com.codecool.navymanager.repository.GunRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.MessageSource;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Locale;
+import java.util.NoSuchElementException;
 
 @Service
 @Transactional(readOnly = true)
 public class GunService {
+    @Autowired
+    MessageSource messageSource;
     private final GunRepository gunRepository;
 
     public GunService(GunRepository gunRepository) {
@@ -21,8 +29,12 @@ public class GunService {
         return gunRepository.findAll().stream().map(GunDto::new).toList();
     }
 
-    public GunDto findById(long id) {
-        return new GunDto(gunRepository.findById(id).orElseThrow());
+    public GunDto findById(long id, Locale locale) {
+        return new GunDto(gunRepository.findById(id)
+                .orElseThrow(() -> new NoSuchElementException(messageSource.getMessage(
+                "no_such",
+                new Object[] {Gun.class.getSimpleName()},
+                locale))));
     }
 
     public List<GunDto> findByCountry(long countryId)  {
@@ -35,16 +47,27 @@ public class GunService {
     }
 
     @Transactional
-    public void update(GunDto gunDto, long id) {
+    public void update(GunDto gunDto, long id, Locale locale) {
         if (gunRepository.existsById(id)) {
             gunRepository.save(gunDto.toEntity());
         } else {
-            throw new IllegalArgumentException("No such gun!");
+            throw new NoSuchElementException(messageSource.getMessage(
+                    "no_such",
+                    new Object[] {Gun.class.getSimpleName()},
+                    locale));
         }
     }
 
     @Transactional
-    public void deleteById(long id) {
-        gunRepository.deleteById(id);
+    public void deleteById(long id, Locale locale) {
+        if (gunRepository.existsById(id)) {
+            gunRepository.deleteById(id);
+        } else {
+            throw  new NoSuchElementException(
+                    messageSource.getMessage(
+                            "no_such",
+                            new Object[] {Gun.class.getSimpleName()},
+                            locale));
+        }
     }
 }
