+++ src/main/java/com/codecool/navymanager/service/RankService.java
@@ -1,15 +1,23 @@
 package com.codecool.navymanager.service;
 
 import com.codecool.navymanager.dto.RankDto;
+import com.codecool.navymanager.entity.Country;
+import com.codecool.navymanager.entity.Rank;
 import com.codecool.navymanager.repository.RankRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.MessageSource;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Locale;
+import java.util.NoSuchElementException;
 
 @Service
 @Transactional(readOnly = true)
 public class RankService {
+    @Autowired
+    MessageSource messageSource;
     private final RankRepository rankRepository;
 
     public RankService(RankRepository rankRepository) {
@@ -22,9 +30,12 @@ public class RankService {
                 .toList();
     }
 
-    public RankDto findById(long id) {
+    public RankDto findById(long id, Locale locale) {
         return new RankDto(rankRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("No such rank!")));
+                .orElseThrow(() -> new NoSuchElementException(messageSource.getMessage(
+                        "no_such",
+                        new Object[] {Rank.class.getSimpleName()},
+                        locale))));
     }
 
     @Transactional
@@ -33,16 +44,27 @@ public class RankService {
     }
 
     @Transactional
-    public void update(RankDto rankDto, long id) {
+    public void update(RankDto rankDto, long id, Locale locale) {
         if (rankRepository.existsById(id)) {
             rankRepository.save(rankDto.toEntity());
         } else {
-            throw new IllegalArgumentException("No such rank!");
+            throw new NoSuchElementException(messageSource.getMessage(
+                    "no_such",
+                    new Object[] {Rank.class.getSimpleName()},
+                    locale));
         }
     }
 
     @Transactional
-    public void deleteById(long id) {
-        rankRepository.deleteById(id);
+    public void deleteById(long id, Locale locale) {
+        if (rankRepository.existsById(id)) {
+            rankRepository.deleteById(id);
+        } else {
+            throw  new NoSuchElementException(
+                    messageSource.getMessage(
+                            "no_such",
+                            new Object[] {Rank.class.getSimpleName()},
+                            locale));
+        }
     }
 }
