+++ src/main/java/com/codecool/navymanager/service/ShipClassService.java
@@ -1,20 +1,26 @@
 package com.codecool.navymanager.service;
 
 
-import com.codecool.navymanager.entity.GunInstallation;
-import com.codecool.navymanager.entity.ShipClass;
+import com.codecool.navymanager.entity.*;
 import com.codecool.navymanager.dto.GunInstallationDto;
 import com.codecool.navymanager.dto.GunDto;
 import com.codecool.navymanager.dto.ShipClassDto;
 import com.codecool.navymanager.repository.ShipClassRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.MessageSource;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Locale;
+import java.util.NoSuchElementException;
 
 @Service
 @Transactional(readOnly = true)
 public class ShipClassService {
+
+    @Autowired
+    MessageSource messageSource;
     private final ShipClassRepository shipClassRepository;
     private final GunService gunService;
 
@@ -42,33 +48,49 @@ public class ShipClassService {
     }
 
     @Transactional
-    public void update(ShipClassDto shipClassDto, long id) {
+    public void update(ShipClassDto shipClassDto, long id, Locale locale) {
         ShipClass oldShipClassData = shipClassRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("No such ship class!"));
+                .orElseThrow(() -> new NoSuchElementException( messageSource.getMessage(
+                        "no_such",
+                        new Object[] {ShipClass.class.getSimpleName()},
+                        locale)));
         ShipClass newShipClassData = shipClassDto.toEntity();
         newShipClassData.setGuns(oldShipClassData.getGuns());
         shipClassRepository.save(newShipClassData);
     }
 
     @Transactional
-    public void deleteById(long id) {
-        shipClassRepository.deleteById(id);
+    public void deleteById(long id, Locale locale) {
+        if (shipClassRepository.existsById(id)) {
+            shipClassRepository.deleteById(id);
+        } else {
+            throw  new NoSuchElementException(
+                    messageSource.getMessage(
+                            "no_such",
+                            new Object[] {ShipClass.class.getSimpleName()},
+                            locale));
+        }
     }
 
     @Transactional
-    public void addGunToShipClass(long shipClassId, GunInstallationDto gunInstallationDto) {
+    public void addGunToShipClass(long shipClassId, GunInstallationDto gunInstallationDto, Locale locale) {
             ShipClass shipClass = shipClassRepository.findById(shipClassId)
-                    .orElseThrow(() -> new IllegalArgumentException("No such ship!"));
-            GunInstallation oldGunInstallation = shipClass.getGuns().stream()
-                    .filter(gunAndQuantity -> gunAndQuantity.getGun().getId().equals(gunInstallationDto.getGun().getId()))
-                    .findAny().orElse(null);
-            if (oldGunInstallation != null) {
-                throw new IllegalArgumentException("That gun was already added to the ship class!");
-            } else {
+                    .orElseThrow(() -> new NoSuchElementException(messageSource.getMessage(
+                            "no_such",
+                            new Object[] {ShipClass.class.getSimpleName()},
+                            locale)));
+            if (shipClass.getGuns().stream()
+                    .noneMatch(gunAndQuantity ->
+                            gunAndQuantity.getGun().getId().equals(gunInstallationDto.getGun().getId()))) {
                 GunInstallation gunInstallation = gunInstallationDto.toEntity();
                 gunInstallation.setShipClass(shipClass);
                 shipClass.getGuns().add(gunInstallation);
                 shipClassRepository.save(shipClass);
+            } else {
+                throw new IllegalArgumentException(messageSource.getMessage(
+                        "invalid_data",
+                        new Object[] {Gun.class.getSimpleName()},
+                        locale));
             }
     }
 
@@ -76,29 +98,45 @@ public class ShipClassService {
     public void updateGunForShipClass(
             long shipClassId,
             long gunId,
-            GunInstallationDto newGunInstallationDto) {
+            GunInstallationDto newGunInstallationDto,
+            Locale locale) {
         ShipClass shipClass = shipClassRepository.findById(shipClassId)
-                .orElseThrow(() -> new IllegalArgumentException("No such ship class!"));
-        GunInstallation gunInstallation = shipClass.getGuns().stream().filter(gQ -> gQ.getGun().getId() == gunId)
-                .findAny().orElseThrow(() -> new IllegalArgumentException("Ship class has no such guns!"));
+                .orElseThrow(() -> new NoSuchElementException(messageSource.getMessage(
+                        "no_such",
+                        new Object[] {Ship.class.getSimpleName()},
+                        locale)));
+        GunInstallation gunInstallation = shipClass.getGuns().stream().filter(gInst -> gInst.getGun().getId() == gunId)
+                .findAny().orElseThrow(() -> new NoSuchElementException(messageSource.getMessage(
+                        "param0_has_no_such_param1",
+                        new Object[] {ShipClass.class.getSimpleName(), Gun.class.getSimpleName()},
+                        locale)));
         gunInstallation.setGun(newGunInstallationDto.getGun().toEntity());
         gunInstallation.setGunQuantity(newGunInstallationDto.getQuantity());
         shipClassRepository.save(shipClass);
     }
 
     @Transactional
-    public void removeGunFromShipClass(long shipClassId, long gunId) {
+    public void removeGunFromShipClass(long shipClassId, long gunId, Locale locale) {
         ShipClass shipClass = shipClassRepository.findById(shipClassId)
-                .orElseThrow(() -> new IllegalArgumentException("No such ship class!"));
-        GunInstallation gunInstallation = shipClass.getGuns().stream().filter(gQ -> gQ.getGun().getId() == gunId)
-                .findAny().orElseThrow(() -> new IllegalArgumentException("Ship class has no such guns!"));
+                .orElseThrow(() -> new NoSuchElementException(messageSource.getMessage(
+                        "no_such",
+                        new Object[] {ShipClass.class.getSimpleName()},
+                        locale)));
+        GunInstallation gunInstallation = shipClass.getGuns().stream().filter(gInst -> gInst.getGun().getId() == gunId)
+                .findAny().orElseThrow(() -> new NoSuchElementException(messageSource.getMessage(
+                        "no_such",
+                        new Object[] {Gun.class.getSimpleName()},
+                        locale)));
         shipClass.getGuns().remove(gunInstallation);
         shipClassRepository.save(shipClass);
     }
 
-    public List<GunDto> findValidGuns(ShipClassDto shipClassDto) {
+    public List<GunDto> findValidGuns(ShipClassDto shipClassDto, Locale locale) {
         var gunIdsOfShipClass = shipClassRepository.findById(shipClassDto.getId())
-                .orElseThrow(() -> new IllegalArgumentException("No such ship class!"))
+                .orElseThrow(() -> new NoSuchElementException(messageSource.getMessage(
+                        "no_such",
+                        new Object[] {ShipClass.class.getSimpleName()},
+                        locale)))
                 .getGuns().stream()
                     .map(gunAndQuantity -> gunAndQuantity.getGun().getId()).toList();
         return gunService.findByCountry(shipClassDto.getCountry().getId()).stream()
